# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CSV Analyser API",
        "description": "High-performance FastAPI backend for CSV data analysis, Open Icecat product enrichment, preview, and visualizations with MongoDB persistence.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/csv/verify": {
            "post": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Verify Csv Dataset",
                "description": "Upload a CSV, audit each product row against live Google/DuckDuckGo web index data,\ncalculate data closeness/fidelity scores, detect discrepancies, and return comprehensive insights.",
                "operationId": "verify_csv_dataset_api_csv_verify_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_verify_csv_dataset_api_csv_verify_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/csv/{file_id}/download-verified": {
            "get": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Download Verified Csv",
                "description": "Download the verified CSV file containing web closeness scores, verification status, and audit insights.",
                "operationId": "download_verified_csv_api_csv__file_id__download_verified_get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/csv/verify/single": {
            "post": {
                "tags": [
                    "CSV Enrichment, Processing & Web Verification"
                ],
                "summary": "Verify Single Product",
                "description": "Audit and cross-verify a single Brand + Model against live Google/DuckDuckGo web index data.",
                "operationId": "verify_single_product_api_csv_verify_single_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SingleVerificationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleVerificationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_verify_csv_dataset_api_csv_verify_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File",
                        "description": "CSV file to cross-verify against live DuckDuckGo web data"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_verify_csv_dataset_api_csv_verify_post"
            },
...
            "DatasetInsights": {
                "properties": {
                    "total_items_checked": {
                        "type": "integer",
                        "title": "Total Items Checked"
                    },
                    "average_closeness_score": {
                        "type": "number",
                        "title": "Average Closeness Score"
                    },
                    "accuracy_grade": {
                        "type": "string",
                        "title": "Accuracy Grade"
                    },
                    "high_confidence_items": {
                        "type": "integer",
                        "title": "High Confidence Items"
                    },
                    "moderate_confidence_items": {
                        "type": "integer",
                        "title": "Moderate Confidence Items"
                    },
                    "discrepancy_items": {
                        "type": "integer",
                        "title": "Discrepancy Items"
                    },
                    "overall_dataset_verdict": {
                        "type": "string",
                        "title": "Overall Dataset Verdict"
                    },
                    "key_takeaways": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Key Takeaways",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "total_items_checked",
                    "average_closeness_score",
                    "accuracy_grade",
                    "high_confidence_items",
                    "moderate_confidence_items",
                    "discrepancy_items",
                    "overall_dataset_verdict"
                ],
                "title": "DatasetInsights"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SingleVerificationRequest": {
                "properties": {
                    "brand": {
                        "type": "string",
                        "title": "Brand",
                        "description": "Brand name, e.g. Samsung, LG, HP"
                    },
                    "model_code": {
                        "type": "string",
                        "title": "Model Code",
                        "description": "Model number or part code, e.g. AR18CY5A, OLED55C3"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "Optional product description/variant, e.g. 1.5 Ton Inverter AC",
                        "default": ""
                    }
                },
                "type": "object",
                "required": [
                    "brand",
                    "model_code"
                ],
                "title": "SingleVerificationRequest"
            },
            "SingleVerificationResponse": {
                "properties": {
                    "brand": {
                        "type": "string",
                        "title": "Brand"
                    },
                    "model_code": {
                        "type": "string",
                        "title": "Model Code"
                    },
                    "web_closeness_score": {
                        "type": "string",
                        "title": "Web Closeness Score"
                    },
                    "closeness_percentage": {
                        "type": "number",
                        "title": "Closeness Percentage"
                    },
                    "verification_status": {
                        "type": "string",
                        "title": "Verification Status"
                    },
                    "verified_market_title": {
                        "type": "string",
                        "title": "Verified Market Title"
                    },
                    "spec_verification_insights": {
                        "type": "string",
                        "title": "Spec Verification Insights"
                    },
                    "live_web_reference": {
                        "type": "string",
                        "title": "Live Web Reference"
                    },
                    "brand_verified": {
                        "type": "boolean",
                        "title": "Brand Verified"
                    },
                    "model_verified": {
                        "type": "boolean",
                        "title": "Model Verified"
                    },
                    "specs_verified": {
                        "type": "boolean",
                        "title": "Specs Verified"
                    },
                    "detected_discrepancies": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Detected Discrepancies",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "brand",
                    "model_code",
                    "web_closeness_score",
                    "closeness_percentage",
                    "verification_status",
                    "verified_market_title",
                    "spec_verification_insights",
                    "live_web_reference",
                    "brand_verified",
                    "model_verified",
                    "specs_verified"
                ],
                "title": "SingleVerificationResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VerificationResponse": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id"
                    },
                    "original_filename": {
                        "type": "string",
                        "title": "Original Filename"
                    },
                    "verified_filename": {
                        "type": "string",
                        "title": "Verified Filename"
                    },
                    "dataset_insights": {
                        "$ref": "#/components/schemas/DatasetInsights"
                    },
                    "preview_rows": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Preview Rows",
                        "default": []
                    },
                    "download_url": {
                        "type": "string",
                        "title": "Download Url"
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "original_filename",
                    "verified_filename",
                    "dataset_insights",
                    "download_url"
                ],
                "title": "VerificationResponse"
            }
        }
    }
}